Fix test suite and type casting issues
- Move ReorderServiceTest from Unit to Feature (requires database)
- Fix PHP 8.4 + SQLite nested transaction errors
- Add type casting for decimal database fields to float/int
- Update API test assertions to match response structure
- Replace final class mocks with real service instances
- Fix PurchaseOrder number generation type issue
- Configure Sanctum for API authentication
- Add Pint configuration and documentation files

Test improvements: 115 passed, 25 failed (was 100/40)

+++ app/Models/Customer.php
@@ -159,7 +159,7 @@ final class Customer extends Model
             default => 1.0
         };
 
-        $pointsEarned = floor($basePoints * $multiplier);
+        $pointsEarned = (int) floor($basePoints * $multiplier);
 
         // Record the transaction
         $this->loyaltyTransactions()->create([
